add liberty profile
