+++ src/main/profiles/liberty/config/server.xml
@@ -2,7 +2,6 @@
 
   <featureManager>
       <feature>microprofile-1.3</feature>
-      <!--feature>usr:opentracingZipkin-0.30</feature-->
   </featureManager>
 
   <httpEndpoint httpPort="${default.http.port}" httpsPort="${default.https.port}"
@@ -11,6 +10,6 @@
   <!-- Automatically expand WAR files and EAR files -->
   <applicationManager autoExpand="true"/>
 
-  <webApplication name="mp-istio-sample-serviceA" location="${app.location}"/>
+  <webApplication name="mp-servicemesh-sample-servicea" location="${app.location}"/>
 
 </server>
